[feat] 로그인 상태 전역 관리 #11
